Basic functionality to generate random tokens and QR codes
- I have NOT tested this in the docker environment
    - The scripts work fine on my laptop in a conda environment
- the two new packages are related to generating QR codes

There is one script to generate random tokens and one to generate QR codes from
the tokens

Note that the scripts are written as scripts with the appropriate argument parsing
These are not checked in anywhere else, so feel free to change the structure if needed.
Hopefully we won't need too many changes, since I do have functions for most of
the actual functionality.

I have put the new files into the `opadmindash` directory - you should be able
to import them once you put them into the PYTHONPATH.

I didn't think that it was a great idea to put them into the `pages` directory,
which is where all the new code lives