Re-load the demographic table on refresh instead of only on dashboard init
This fixes https://github.com/e-mission/e-mission-docs/issues/1000
Per https://github.com/e-mission/e-mission-docs/issues/1000#issuecomment-1749910030

Before this fix, while we had `update_store_demographics`, just like
`update_store_uuids`, we create the demographic_data once and pass it in. while
we initialize the UUIDs to empty and update them when the data, start_date or
end_date change. Since there was no callback for the demographic data, it is
treated as a static dataset. The fix was as simple as initializing a callback
for it as well.