use unambiguous format on datepicker (D MMM Y)
Consider "1/5/2024" – in the US this seems like January 5, while in most of the world it seems like May 1. Whether we did D/M/Y or M/D/Y, it'd probably cause some confusion since the project is used on an international scale.

So instead of just digits, the new format will be "5 Jan 2024". Unambiguous and still pretty concise.