Skip sorting while processing trips group
The process_trip_group function,  processes a group of trips, assigns color to each group, sorts the trips within each group and stores the processed data in a dictionary.

We are skipping sorting part as we are plotting all co-ordinates collectively and there is no advantage of sorting it here.