make datepicker's "today" recompute on page load
fixes the issue described in https://github.com/e-mission/op-admin-dashboard/pull/96#issuecomment-1919816848

To recompute a new date on refresh, "layout" needs to be a function. This required reorganizing the page contents.
Some things can stay static but the things that need to be recomputed are reorganized into "make_" functions.

Inside make_controls() is where we generate the dates, and this gets called on every refresh.

The resulting DOM is the same.