auto-adjust zoom on map
When the map was updated, it would just zoom in on whatever was the first location point, which makes for an annoying UX if you are trying to look at a lot of data spread over a large area.

These changes set the view of the map to the bounding box of the points being shown.
This isn't supported natively by dash so there's a new function, thanks to folks on plotly forum