overhaul the 'tokens' page
The tokens page was not fully functional, behaved unexpectedly, and was not very user-friendly. I rewrote it to address this.

Moved the controls for token generation into a modal accessible by clicking a "Generate more tokens" button.
Autofil value of 'program' input with the value of STUDY_CONFIG
Instead of 'For testing' checkbox, use dropdown for 'subgroup' (which will always include 'test' as an option)
Remove generate_random_tokens.py, use new functions from emcommon.auth.opcode instead

Used Dash Bootstrap components for new modal + buttons, upgraded table to Dash Ag Grid

Removed the id column and added the in_use column which informs the user whether any entries in the UUIDs db are using that token. For admins looking to assign tokens to new users, this feature makes it easy to ensure that no opcode is given to multiple users

In programs with more than about 50 opcodes, it is very slow to create all the QR code images at the same time. I decided to have QR codes be created on demand, so the values of the qr_code column are "(click to reveal)" until clicked, then the image is created + displayed

Instead of saving '.png' files to an arbitrary location on the filesystem, image generation is now in base64 string format. It also does not require pil anymore

Exporting to .zip still works

Exporting the table still works