fix humanized/raw values toggle
This callback was made for the plotly dash datatable and no longer works with Dash Ag Grid.
Instead of just setting hidden_column property, we need to adjust the columnDefs and set the 'hide' property per column def.

Updated the ID of the trips table since it is now a dict `{'type': 'data_table', 'id': 'trips'}`, not a string `'trips-datatable'`
Also adjusted the `table_id` parameter to `populate_datatable` to ensure that each tab's datatable has a unique ID