store merged UUIDs with Profiles after loading them upfront
We have been querying for user profiles on both the overview (i.e. initialization of the dashboard), and again on the data page for the UUIDs table, which is redundant.
We may as well query all of the Profiles alongside the UUIDs and store them together.
So in db_utils, instead of query_uuids, we'll have query_users which gets both the UUIDs and the Profiles, merges them, and returns a "users" dataframe

We no longer need add_user_stats because we don't have to query each user separately; a few columns just need to be renamed before we display the UUIDs table, which we do in data.py render_content

I also changed the display name of some columns (e.g. app_version -> client_app_version) because the longer names reflect what's actually stored in the DB. I think it's simpler and clearer to just match the data model here; it's not like client_app_version is a terribly long and jargony name compared to app_version