fix the ordering of columns on data tables
In Python, sets are unique and unordered.
Several of the functions in permissions.py used sets to track the columns that show up, which ensures no duplicates.
However, it causes inconsistent ordering of the columns and is bad UX.
Changing these to lists and only adding columns that aren't already added gives an ordered list without duplicates.

With the order now consistent, I also took this opportunity to ensure that user_token and user_id appear first on any tables that they are present