make mode dropdowns generic, add colors for sensed mode, refactor
Instead of defining separate functions for creating the dropdowns and doing the filtering based on the selected values, we can define generic functions (filter_trips and create_modes_dropdown_options) that have a `trip_key` or `mode_key` parameter, which can be 'mode_confirm' for labeled modes, 'data.primary_sensed_mode' for sensed modes, or 'data.primary_ble_sensed_mode' for BLE mode.
For the trip lines map, we can also use this to filter by user with a trip_key of 'user_id'.

When it was just labeled mode, we sorted all the trips from store-trips and organized them by user_id and mode, and we put this in store-trips-map. Now with multiple mode filters, it makes less sense to precompute all these groupings. Instead we'll do this sorting as needed while generating the figures. This means store-trips-map is gone and the code actually becomes a lot simpler

Made all maps use the color of the primary mode

Extracted the hover text into a separate function

Made create_lines_map match the structure of create_heatmap_fig and create_bubble_fig

Added indicator for # of trips shown on map; if 0 trips do not show figure