🚩 💩 Disable token generation temporarily
The admin dashboard currently is primarily is a data/viz dashboard, so it
accesses the database read-only. This is the first feature that would write to
the database, on the way towards a more "back office" type of operation.

We will re-enable this once we get internal buy-in at NREL on the shift in focus

@JGreenlee for visibility